Convert some recent tests to JUnit 4
