--- src/test/groovy/bugs/Groovy8964Bug.groovy
@@ -1,43 +1,0 @@
-/*
- *  Licensed to the Apache Software Foundation (ASF) under one
- *  or more contributor license agreements.  See the NOTICE file
- *  distributed with this work for additional information
- *  regarding copyright ownership.  The ASF licenses this file
- *  to you under the Apache License, Version 2.0 (the
- *  "License"); you may not use this file except in compliance
- *  with the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing,
- *  software distributed under the License is distributed on an
- *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- *  KIND, either express or implied.  See the License for the
- *  specific language governing permissions and limitations
- *  under the License.
- */
-package groovy.bugs
-
-class Groovy8964Bug extends GroovyTestCase {
-    void testInstanceVarargMethodNotMaskedByStaticMethodWithSameNumberOfArgs() {
-        assertScript '''
-            class Example {
-
-                def method(String... args) {
-                    'vararg'
-                }
-
-                static method(List<String> args, File workDirectory, Appendable out, Appendable err) {
-                    'multi'
-                }
-
-                def execute() {
-                    method("a", "b", "c", "d")
-                }
-            }
-
-            Example ex = new Example()
-            assert ex.execute() == 'vararg'
-        '''
-    }
-}
