+++ src/test/groovy/bugs/Groovy9007.groovy
@@ -1,24 +1,30 @@
 /*
- *  Licensed to the Apache Software Foundation (ASF) under one
- *  or more contributor license agreements.  See the NOTICE file
- *  distributed with this work for additional information
- *  regarding copyright ownership.  The ASF licenses this file
- *  to you under the Apache License, Version 2.0 (the
- *  "License"); you may not use this file except in compliance
- *  with the License.  You may obtain a copy of the License at
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
  *
- *    http://www.apache.org/licenses/LICENSE-2.0
+ *	http://www.apache.org/licenses/LICENSE-2.0
  *
- *  Unless required by applicable law or agreed to in writing,
- *  software distributed under the License is distributed on an
- *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- *  KIND, either express or implied.  See the License for the
- *  specific language governing permissions and limitations
- *  under the License.
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
  */
 package groovy.bugs
 
-class Groovy9007Bug extends GroovyTestCase {
+import org.junit.Test
+
+import static groovy.test.GroovyAssert.assertScript
+
+final class Groovy9007 {
+
+    @Test
     void testProtectedFieldInClosureInEnum() {
         assertScript '''
             @groovy.transform.CompileStatic
@@ -35,6 +41,7 @@ class Groovy9007Bug extends GroovyTestCase {
         '''
     }
 
+    @Test
     void testPrivateFieldInClosureInEnum() {
         assertScript '''
             @groovy.transform.CompileStatic
@@ -51,7 +58,7 @@ class Groovy9007Bug extends GroovyTestCase {
         '''
     }
 
-    // GROOVY-8978
+    @Test // GROOVY-8978
     void testProtectedFieldInChildWithExplicitThis() {
         assertScript '''
             import groovy.transform.CompileStatic
