+++ src/test/groovy/bugs/Groovy9010.groovy
@@ -7,7 +7,7 @@
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
  *
- * 	http://www.apache.org/licenses/LICENSE-2.0
+ *	http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
@@ -19,29 +19,18 @@
 package groovy.bugs
 
 import groovy.transform.CompileStatic
-import org.codehaus.groovy.control.CompilerConfiguration
 import org.junit.Test
 
-import static org.codehaus.groovy.control.ParserPluginFactory.antlr2
+import static groovy.test.GroovyAssert.assertScript
 
 @CompileStatic
-final class Groovy9184 {
+final class Groovy9010 {
 
-    @Test(timeout=1500L)
-    void testEnumWithinEnum() {
-        def config = new CompilerConfiguration(pluginFactory: antlr2())
-
-        new GroovyShell(config).evaluate('''
-            enum Outer {
-                A, B
-                enum Inner {
-                    X, Y
-                }
-            }
-            assert Outer.A.name() == 'A'
-            assert Outer.A.ordinal() == 0
-            assert Outer.Inner.Y.name() == 'Y'
-            assert Outer.Inner.Y.ordinal() == 1
-        ''')
+    @Test
+    void testASTTestWithNoSpacesInClosure() {
+        assertScript '''
+            @groovy.transform.ASTTest({assert 1==1})
+            def x
+        '''
     }
 }
