+++ src/test/groovy/bugs/Groovy9115.groovy
@@ -7,7 +7,7 @@
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
  *
- * 	http://www.apache.org/licenses/LICENSE-2.0
+ *	http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
@@ -19,22 +19,48 @@
 package groovy.bugs
 
 import groovy.transform.CompileStatic
-import org.codehaus.groovy.ast.expr.ConstantExpression
 import org.junit.Test
 
 import static groovy.test.GroovyAssert.assertScript
 
 @CompileStatic
-final class Groovy9153 {
+final class Groovy9115 {
 
     @Test
-    void testConstantExpressionNULL() {
+    void testSetPropertyInIfStmt() {
         assertScript '''
             @groovy.transform.CompileStatic
-            File tempDir() {
-                return File.createTempDir()
+            class Derived {
+                def m() {
+                    if (true) {
+                        File file = File.createTempFile("hello${System.nanoTime()}", ".tmp")
+                        file.text = 'Groovy9115Bug'
+                        assert 'Groovy9115Bug' == file.text
+                    }
+
+                    return null
+                }
             }
+
+            new Derived().m()
         '''
-        assert ConstantExpression.NULL.getNodeMetaData().isEmpty()
+    }
+
+    @Test
+    void testSetProperty() {
+        assertScript '''
+            @groovy.transform.CompileStatic
+            class Derived {
+                def m() {
+                    File file = File.createTempFile("hello${System.nanoTime()}", ".tmp")
+                    file.text = 'Groovy9115Bug'
+                    assert 'Groovy9115Bug' == file.text
+
+                    return null
+                }
+            }
+
+            new Derived().m()
+        '''
     }
 }
